Update Autonomous RC Car platform
Hi! So the antennas will be used in initial testing to make sure I can control the car using an Xbox controller. 
I'll know that the RC car is compatible if I can disconnect the receiver from the speed controller/motor. This is important because I need the PCA servo driver to integrate with the car.
The wires are used to bridge the gap between the Nvidia Jetson and the servo driver(they're the breadboard).