Update project focus to modern NES Zapper
In this commit, we are updating the focus of our project from a modern version of the classic NES game, Duck Hunt, to a modern version of the NES Zapper. The Duck Hunt game will now be treated as a demo project, while the main project will be the modern NES Zapper.

We have revised the project description to reflect this change and to provide more details about the features and functionality of the modern NES Zapper.

This commit also includes updates to the project documentation and any necessary code changes to reflect the revised project focus.