ARDUINO AMPHIBIOUS ROVER
The rover will be built using an Arduino microcontroller, a variety of sensors, and waterproof motors and hardware to enable it to move through water. I plan to use readily available components, however the entire project will be new and original, leaving myself to experiment with new concepts rather than following a strict tutorial.