Update Val171.md
Well, encountered a major hurdle, the website I was ordering the chassis from turned out to be a scam so I'll be making the chassis out of balsa wood and cardboard. The electronics remain more or less the same. I am leaving a buffer incase I need more parts. A lot of these items are currently on discount thus prices may change in the next few days depending on how long it takes for this to get approved.